adk-study-planner/requirements.txt
google-adk>=1.19.0
arxiv>=2.1.0
python-dotenv>=1.0.0